requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.32
asyncpg==0.30.0
geoalchemy2==0.14.1
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.10
